requirements.txt
absl-py==0.8.1
altair==3.2.0
argh==0.26.2
astor==0.8.0
attrs==19.3.0
base58==1.0.3
bert-serving-client==1.9.8
bert-serving-server==1.9.8
blinker==1.4
boto3==1.10.2
botocore==1.13.2
certifi==2019.9.11
chardet==3.0.4
Click==7.0
decorator==4.4.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
faiss==1.5.3
future==0.18.1
gast==0.2.2
google-pasta==0.1.7
GPUtil==1.4.0
grpcio==1.24.3
h5py==2.10.0
idna==2.8
importlib-metadata==0.23
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==3.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
numpy==1.17.3
opt-einsum==3.1.0
pandas==0.25.2
pathtools==0.1.2
Pillow==6.2.1
protobuf==3.10.0
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
streamlit==0.49.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
tzlocal==2.0.0
urllib3==1.25.6
validators==0.14.0
watchdog==0.9.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
